pyproject.toml
[tool.poetry]
name = "pyarango-migrations"
version = "0.1.0"
description = "Python migration utilities for ArangoDB"
authors = ["John Santiago Jr. <jdsantiagojr@protonmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/FiniteStateInc/pyarango-migrations"

packages = [{ include = "toolbox" }]

[tool.poetry.scripts]
toolbox = 'bin.cli:main'

[tool.poetry.dependencies]
python = "^3.10"
python-arango = "^7.6.1"
click = "^8.1.7"

[tool.black]
line-length = 128
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 128
atomic = true

[tool.flake8]
# E221: Multiple spaces before operator
# E402: Module level import not at top of file
# E501: Line too long
# W503: Line break occurred before a binary operator
ignore = ["E221", "E402", "E501", "W503"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
